used combined reducers
