+++ smurfs/src/reducers/index.js
@@ -1,48 +1,6 @@
-/*
-  Be sure to import in all of the action types from `../actions`
-*/
-import { FETCHED, ERRORS, FETCHING } from '../actions';
+import { combineReducers } from 'redux';
+import { smurfReducer } from './smurfReducer';
 
-/*
- Your initial/default state for this project could *Although does not have to* look a lot like this
- {
-   smurfs: [],
-   fetchingSmurfs: false
-   addingSmurf: false
-   updatingSmurf: false
-   deletingSmurf: false
-   error: null
- }
-*/
-const initialState = { smurfs: [], smurf: {}, fetching: false, error: '' };
-
-export default (state = initialState, action) => {
-  switch (action.type) {
-    case FETCHING:
-      return { ...state, fetching: true };
-
-    case FETCHED:
-      return {
-        ...state,
-        smurfs: action.payload,
-        fetching: false
-      };
-
-    case ERRORS:
-      return {
-        ...state,
-        fetching: false,
-        error: `${action.payload}`
-      };
-
-    default:
-      return state;
-  }
-};
-/*
-  You'll only need one smurf reducer for this project.
-  Feel free to export it as a default and import as rootReducer. 
-  This will guard your namespacing issues.
-  There is no need for 'combineReducers' in this project.
-  Components can then read your store as, `state` and not `state.fooReducer`.
-*/
+export default combineReducers({
+  smurfReducer
+});
